Extract grants impl
