+++ pkg/sdk/grants.go
@@ -15,12 +15,6 @@ type Grants interface {
 	Show(ctx context.Context, opts *ShowGrantOptions) ([]*Grant, error)
 }
 
-var _ Grants = (*grants)(nil)
-
-type grants struct {
-	client *Client
-}
-
 type Grant struct {
 	CreatedOn   time.Time
 	Privilege   string
@@ -129,27 +123,6 @@ type GrantOnSchemaObjectIn struct {
 	InSchema         *DatabaseObjectIdentifier `ddl:"identifier" sql:"IN SCHEMA"`
 }
 
-func (v *grants) GrantPrivilegesToAccountRole(ctx context.Context, privileges *AccountRoleGrantPrivileges, on *AccountRoleGrantOn, role AccountObjectIdentifier, opts *GrantPrivilegesToAccountRoleOptions) error {
-	if opts == nil {
-		opts = &GrantPrivilegesToAccountRoleOptions{}
-	}
-	opts.privileges = privileges
-	opts.on = on
-	opts.accountRole = role
-	if err := opts.validate(); err != nil {
-		return err
-	}
-	sql, err := structToSQL(opts)
-	if err != nil {
-		return err
-	}
-	_, err = v.client.exec(ctx, sql)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
 type RevokePrivilegesFromAccountRoleOptions struct {
 	revoke         bool                        `ddl:"static" sql:"REVOKE"`
 	GrantOptionFor *bool                       `ddl:"keyword" sql:"GRANT OPTION FOR"`
@@ -158,27 +131,6 @@ type RevokePrivilegesFromAccountRoleOptions struct {
 	accountRole    AccountObjectIdentifier     `ddl:"identifier" sql:"FROM ROLE"`
 	Restrict       *bool                       `ddl:"keyword" sql:"RESTRICT"`
 	Cascade        *bool                       `ddl:"keyword" sql:"CASCADE"`
-}
-
-func (v *grants) RevokePrivilegesFromAccountRole(ctx context.Context, privileges *AccountRoleGrantPrivileges, on *AccountRoleGrantOn, role AccountObjectIdentifier, opts *RevokePrivilegesFromAccountRoleOptions) error {
-	if opts == nil {
-		opts = &RevokePrivilegesFromAccountRoleOptions{}
-	}
-	opts.privileges = privileges
-	opts.on = on
-	opts.accountRole = role
-	if err := opts.validate(); err != nil {
-		return err
-	}
-	sql, err := structToSQL(opts)
-	if err != nil {
-		return err
-	}
-	_, err = v.client.exec(ctx, sql)
-	if err != nil {
-		return err
-	}
-	return nil
 }
 
 type grantPrivilegeToShareOptions struct {
@@ -201,23 +153,6 @@ type OnTable struct {
 	AllInSchema DatabaseObjectIdentifier `ddl:"identifier" sql:"ALL TABLES IN SCHEMA"`
 }
 
-func (v *grants) GrantPrivilegeToShare(ctx context.Context, privilege ObjectPrivilege, on *GrantPrivilegeToShareOn, to AccountObjectIdentifier) error {
-	opts := &grantPrivilegeToShareOptions{
-		privilege: privilege,
-		On:        on,
-		to:        to,
-	}
-	if err := opts.validate(); err != nil {
-		return err
-	}
-	sql, err := structToSQL(opts)
-	if err != nil {
-		return err
-	}
-	_, err = v.client.exec(ctx, sql)
-	return err
-}
-
 type revokePrivilegeFromShareOptions struct {
 	revoke    bool                        `ddl:"static" sql:"REVOKE"`
 	privilege ObjectPrivilege             `ddl:"keyword"`
@@ -235,23 +170,6 @@ type RevokePrivilegeFromShareOn struct {
 type OnView struct {
 	Name        SchemaObjectIdentifier   `ddl:"identifier" sql:"VIEW"`
 	AllInSchema DatabaseObjectIdentifier `ddl:"identifier" sql:"ALL VIEWS IN SCHEMA"`
-}
-
-func (v *grants) RevokePrivilegeFromShare(ctx context.Context, privilege ObjectPrivilege, on *RevokePrivilegeFromShareOn, id AccountObjectIdentifier) error {
-	opts := &revokePrivilegeFromShareOptions{
-		privilege: privilege,
-		On:        on,
-		from:      id,
-	}
-	if err := opts.validate(); err != nil {
-		return err
-	}
-	sql, err := structToSQL(opts)
-	if err != nil {
-		return err
-	}
-	_, err = v.client.exec(ctx, sql)
-	return err
 }
 
 type ShowGrantOptions struct {
@@ -283,31 +201,4 @@ type ShowGrantsTo struct {
 type ShowGrantsOf struct {
 	Role  AccountObjectIdentifier `ddl:"identifier" sql:"ROLE"`
 	Share AccountObjectIdentifier `ddl:"identifier" sql:"SHARE"`
-}
-
-func (v *grants) Show(ctx context.Context, opts *ShowGrantOptions) ([]*Grant, error) {
-	if opts == nil {
-		opts = &ShowGrantOptions{}
-	}
-	if err := opts.validate(); err != nil {
-		return nil, err
-	}
-	sql, err := structToSQL(opts)
-	if err != nil {
-		return nil, err
-	}
-	var rows []grantRow
-	err = v.client.query(ctx, &rows, sql)
-	if err != nil {
-		return nil, err
-	}
-	grants := make([]*Grant, 0, len(rows))
-	for _, row := range rows {
-		grant, err := row.toGrant()
-		if err != nil {
-			return nil, err
-		}
-		grants = append(grants, grant)
-	}
-	return grants, nil
 }
